% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AffinityNetworkFusion.R
\name{ANF}
\alias{ANF}
\title{Fuse affinity networks through one-step or two-step random walk}
\usage{
ANF(Wall, K = 20, weight = NULL, type = 2, alpha = c(1, 1, 0, 0, 0, 0,
  0, 0), verbose = FALSE)
}
\arguments{
\item{Wall}{a list of affinity matrices of the same shape}

\item{K}{the number of k nearest neighbors for function kNN_graph}

\item{weight}{the weight for each view}

\item{type}{1: one-step random walk; 2: two-step random walk}

\item{alpha}{used as weights for terms in type 2 (two-step random walk)}

\item{verbose}{Boolean; if true, print some information}
}
\value{
a fused transition matrix (representing fused network)
}
\description{
Fuse affinity networks through one-step or two-step random walk
}
\examples{
D1 = matrix(runif(400),20)
W1 = affinity_matrix(D1, 5)
D2 = matrix(runif(400),20)
W2 = affinity_matrix(D1, 5)
W = ANF(list(W1, W2), K=10)
}
