% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectralClustering.R
\name{eval_clu}
\alias{eval_clu}
\title{Evaluate cluster result}
\usage{
eval_clu(true_class, w = NULL, d = NULL, k = 10, num_clu = NULL,
  surv = NULL, type_L = 2, verbose = T)
}
\arguments{
\item{true_class}{A named vector of true class labels}

\item{w}{affinity matrix}

\item{d}{distance matrix; 
if w is NULL, calcuate w using d}

\item{k}{an integer; if w is null, w = affinity_matrix(d, k)}

\item{num_clu}{an integer; number of clusters; 
if NULL, set num_clu to be the number of classes using true_class}

\item{surv}{a data.frame with at least two columns: 
time (days_to_death or days_to_last_follow_up), and censored (Boolean)}

\item{type_L}{an integer of 1, 2, or 3; choice of normalized graph Laplacian}

\item{verbose}{Boolean; if true, print some information}
}
\value{
a vector of Normalized Mutual Informatin (NMI), Adjusted Rand Index (ARI),
and -log10(p) of log rank test of survival distributions of patient clusters
}
\description{
Evaluate cluster result
}
