% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AffinityNetworkFusion.R
\name{kNN_graph}
\alias{kNN_graph}
\title{Calculate k-nearest-neighbor graph from affinity matrix
normalize it as transition matrix}
\usage{
kNN_graph(W, K)
}
\arguments{
\item{W}{affinity matrix}

\item{K}{the number of k nearest neighbors}
}
\value{
an transition matrix of the same shape as W
}
\description{
Calculate k-nearest-neighbor graph from affinity matrix
normalize it as transition matrix
}
\examples{
D = matrix(runif(400),20)
W = affinity_matrix(D, 5)
S = kNN_graph(W, 5)
}
