% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectralClustering.R
\name{pod}
\alias{pod}
\title{Finding optimal discrete solutions for spectral clustering}
\usage{
pod(Y, verbose = FALSE)
}
\arguments{
\item{Y}{first k eigenvectors of graph Laplacian corresponding to k smallest eigenvalues}

\item{verbose}{Boolean; if true, print some information}
}
\value{
class assignment matrix (0-1)
}
\description{
Finding optimal discrete solutions for spectral clustering
}
\examples{
D = matrix(runif(400),20)
A = affinity_matrix(D, 5)
d = rowSums(A)
L = diag(d) - A
NL = diag(1/d) \%*\% L
e = eigen(NL)
Y = Re(e$vectors[,-1:-17])
X = pod(Y)
}
\references{
Stella, X. Yu, and Jianbo Shi. "Multiclass spectral clustering." ICCV. IEEE, 2003.
}
