% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectralClustering.R
\name{spectral_clustering}
\alias{spectral_clustering}
\title{spectral_clustering}
\usage{
spectral_clustering(A, k, type = 2, verbose = FALSE)
}
\arguments{
\item{A}{affinity matrix}

\item{k}{the number of clusters}

\item{type}{choice of normalized graph Laplacian}

\item{verbose}{Boolean; if true, print user-friendly information}
}
\value{
a numeric vector as class labels
}
\description{
spectral_clustering
}
\examples{
D = matrix(runif(400),20)
A = affinity_matrix(D, 5)
labels = spectral_clustering(A, 2)
}
